requirements.txt
Django==1.7.2
MySQL-python==1.2.5
Pillow==2.7.0
argparse==1.2.1
django-grappelli==2.6.3
django-widget-tweaks==1.3
psycopg2==2.5.4
wsgiref==0.1.2
nonude==1.1
